+++ packages/visual-grid-client/test/e2e/openEyes.e2e.test.js
@@ -90,7 +90,6 @@ describe('openEyes', () => {
         {deviceName: 'iPhone X'},
       ],
       showLogs: process.env.APPLITOOLS_SHOW_LOGS,
-      saveDebugData: process.env.APPLITOOLS_SAVE_DEBUG_DATA,
     })
 
     const scriptHooks = {
@@ -195,7 +194,6 @@ describe('openEyes', () => {
         {deviceName: 'iPhone X'},
       ],
       showLogs: process.env.APPLITOOLS_SHOW_LOGS,
-      saveDebugData: process.env.APPLITOOLS_SAVE_DEBUG_DATA,
     })
 
     const scriptHooks = {